% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agetimes}
\alias{agetimes}
\title{Counts of Fatal Automobile Accidents in the United States, 2004-2018 by Age of victim and time of day of accident}
\format{A tibble with 1875 rows and 4 columns

\describe{
\item{\code{age}}{Age group of victim}
\item{\code{time}}{Time of day (hour). Character variable.}
\item{\code{time_fct}}{Time of day (hour). Ordered factor; 'Unknown Hours' coded as NA.}
\item{\code{year}}{Year}
\item{\code{n}}{Count}
}}
\source{
Fatality Analysis Reporting System (FARS): 2004-2017 Final File and 2018 Annual Report File (ARF)
}
\usage{
agetimes
}
\description{
A tibble containing selected age- and time-level characteristics of automobile crashes that resulted in at least one fatality, 2004-2018
}
\keyword{datasets}
