% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{FARS2013}
\alias{FARS2013}
\title{Fatality Analysis Reporting System data for 2013}
\format{A tibble data frame with 30202 rows and 50 variables including
\describe{
\item{STATE}{A numerical value for the state or territory of the US in which the fatality occured.
In alphabetical order, includes Puerto Rico, DC and the Virgin islands (numbers 3,7 and 14 removed)}
\item{VE_TOTAL}{Number of vehicles involved in the crash}
\item{PERSONS}{Number of people involved in the crash}
\item{NO_LANES}{Number of lanes on the road on which the crash occured}
\item{SP_LIMIT}{Speed limit at the crash location}
\item{FATALS}{Number of people who died in the crash}
\item{DRUNK_DR}{Number of drivers influenced by alcohol in the crash}
\item{...}{}
}}
\source{
Publication from the National Highway Traffic Safety Administration to provide an overall measure of highway safety.
\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\usage{
FARS2013
}
\description{
Fatality Analysis Reporting System data for 2013
}
\keyword{datasets}
