% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the file name which will be reading by the function}
}
\value{
This function returns a dataframe with the information in the file. If the file does not
exist, it will return an error.
}
\description{
This is a simple function that reads data from a .csv file. That file is coming from the US
National Highway Traffic Safety Administration's Fatality Analysis Reporting System, which is
a nationwide census providing the American public yearly data regarding fatal injuries suffered
in motor vehicle traffic crashes.
}
\examples{
filename_param <- system.file('data', 'accident_2013.csv.bz2',
                             package = 'farsdata')
fars_read(filename = filename_param)

}
