% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector with a list of years}
}
\value{
This function returns a dataframe with the months with data by year. If any of the years do not
exist, it will return NULL and a warning.
}
\description{
This is a function that reads the months with data by year.
}
\examples{
fars_read_years(2013)
fars_read_years(c(2013,2014))

}
\seealso{
\link{fars_read}

\link{make_filename}
}
