% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string or an integer giving the year}
}
\value{
This function returns a character string with the file name.
}
\description{
This is a simple function that makes a file name.
}
\examples{
make_filename("2013")
make_filename("2015")

}
