% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{persons}
\alias{persons}
\title{Person-level counts of Fatal Automobile Accidents in the United States, 2004-2018, by sex and race of victim}
\format{A tibble with 405 rows and 4 columns

\describe{
\item{\code{sex}}{Sex of victim}
\item{\code{race}}{Race/ethcnicity of victim}
\item{\code{year}}{Year}
\item{\code{n}}{Count}
}}
\source{
Fatality Analysis Reporting System (FARS): 2004-2017 Final File and 2018 Annual Report File (ARF)
}
\usage{
persons
}
\description{
A tibble containing selected person-level characteristics of automobile crashes that resulted in at least one fatality, 2004-2018
}
\keyword{datasets}
