% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create map of Fatality Analysis Reporting System data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Number of state}

\item{year}{A year of the data to plot}
}
\value{
The function returns a plot with map of a state and the accidents place from the selected year.
}
\description{
The function plots the data from a selected year and state on a map.
}
\note{
The files should be named 'accident_year.csv.bz2' otherwise the function will return an error.
}
\examples{
fars_map_state(1, 2013)

}
