% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display accidents map by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An Integer with the State Code}

\item{year}{A string or an integer with the input year}
}
\value{
None
}
\description{
It displays a plot with a state map which includes the accidents location by year
If the state number is incorrect then it shows an error
}
\examples{
\dontrun{ fars_map_state(45, 2014)}


}
