% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a map of accidents by State and Year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number corresponding to a State in the FARS data.}

\item{year}{A character or number representing the target year.}
}
\value{
A plot of the accident locations in the State in the target year
}
\description{
Draw the location of each accident in a given State in a given year.
}
\details{
Loads the file with the name \code{"accidnet_year.csv.bz2"} where year
corresponds to the value of \code{year}.
Calls map::maps and graphics::points to plot the accidnets by latitude and longtude.
Will produce an error if the \code{state.num} is incorrect or the file for
the given year does not exist.
}
\examples{
\dontrun{
fars_map_state(17, 2016)
fars_map_state(21, "2016")
}
}

