% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsdata.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot state accident data for year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the numerical identifier of the state to plot.}

\item{year}{the year to process and display.}
}
\description{
Plot a map of a state with points for each
case of the given year.
}
\examples{
fars_map_state(1, 2013)

}
