% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state number that corresponds to a state in the FARS dataset}

\item{year}{the year in question}
}
\value{
a plot of the fatalities in a particular state in
    particular year
}
\description{
returns a plot of the fatalities in a particular state in
    particular year
}
\details{
The function plots a data point for each fatality in the dataset for the particular state

    The LONGITUDE and LATITUDE values are cleaned up before being passed to the map routine
    with invalid entires being marked as NA.  The map routine then removes NA values

    The function will return an error if the state.num is not in the dataset
    It will also return a message if there are no value
}
\examples{
\dontrun{
fars_map_state(1,2013)
fars_map_state(5,2014)
}

}
