% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Function to plot the crashes' locations on a US state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{is an numerical scalar indicating which US state has to be considered}

\item{year}{is a numeric scalar specifying which yearly data
(regarding fatal injuries suffered in motor vehicle traffic crashes)
have to be loaded}
}
\value{
NULL. A plot of the US state with the locations of the accidents is printed but not returned.
}
\description{
This function loads the fars data, checks that whether US state number is valid,
checks whether there are crashes to plot, checks whether the latitude and longitudes are
correct, and finally plot the locations of the accidents
using the US state map where they took place.
}
\details{
This function loads the fars data, checks that whether US state number is valid (otherwise it stops),
checks whether there are crashes to plot (otherwise it sends the message
"no accidents to plot"), checks whether the latitude and longitudes are
correct (otherwise it inserts na), and finally plots the locations of the accidents
using the US state map where they took place.
}
\examples{

 data_file<-system.file("extdata","accident_2013.csv.bz2", package = "farsdata")
 setwd(dirname(data_file))
 library(maps)
 fars_map_state(48, 2014)

}

