% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{plotting the location of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer which indicates one state of US}

\item{year}{an integer indicates the year of the fars file}
}
\value{
NULL, just plot the points of the fars data accident
}
\description{
This function read the fars data of a specific year and state then and then try to plot the
  location of the accident in that state and year.
}
\examples{
\dontrun{
fars_summarize_years(33 , 2013)
}

}
