% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number associated with the state of interest based on the state numbers in the original data}

\item{year}{A four-digit year representing the year of data of interest}
}
\value{
map of accident locations for the selected state and year
}
\description{
fars_map_state takes a state ID number and year to create a map of accident locations
}
\details{
This function uses a state ID and four-digit year and generates a map of accident locations for the specified state
within the specified year.
}
\examples{
\dontrun{
fars_map_state(25, 2013)
}
}
\seealso{
\link{make_filename}

\link{fars_read}
}
