% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Title: fars_map_state
The function fars_map_state accepts a state number and a year as arguments. With make_filename function it creates the filename and with fars_read it obtains the corresponding data frame.
It gives an error if the state number does not exist. Otherwise it creates plots based on latitude and longitude, using maps and graphics.
Uses make_filename(year)
fars_read(filename)}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{(state number)}

\item{year}{(the year from which we want the data file)}
}
\value{
a dataframe with the longitud and the latitud of the introduced state and year.
}
\description{
Title: fars_map_state
The function fars_map_state accepts a state number and a year as arguments. With make_filename function it creates the filename and with fars_read it obtains the corresponding data frame.
It gives an error if the state number does not exist. Otherwise it creates plots based on latitude and longitude, using maps and graphics.
Uses make_filename(year)
fars_read(filename)
}
\examples{
\dontrun{fars_map_state(3,2013)}

}
