% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{----5-----
fars_map_state: Creates a visualisation of the accidents on a US state map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The ID of a US state as specified in the FARS data pack.}

\item{year}{The selected year the visualisation will be based on.}
}
\value{
Returns a graph of the accidents based on the \code{state.num} and
        \code{year} inputs. Returns an error if the state and/ or year do not exist in the
        data set.
}
\description{
The function accepts a state ID and year and returns a map visualisation of the accidents.
             The state ID has to be an integer or numerical.
}

