% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS Data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number (integer, numeric or string) that can be coerced into a string,
between 1 and 56 (excluding3,7 and 14), representing alphabetically the 50 us states,
DC, Puerto rico, the Virgin islands.}

\item{year}{A number (integer, numeric or string) that can be coerced into a string to
represent the year of the data, can be any number.}
}
\value{
Returns a map of fatalities in the given state for the given year,
 fatality sites are indicated in black.
}
\description{
\code{fars_map_state} maps Fatality Analysis Reporting System data for a given state and year,
 plotting individual vehicle fatality sites over a map of the state for a given year. Requires valid FARS csv in the WD.
}
\note{
If the state number provided is outside of the range of possible states and territories returns:
 "invalid STATE number: "\code{state.num}.

If there are no fatalities found, returns:
 no accidents to plot"
}
\examples{
\dontrun{fars_map_state(1,2013)
fars_map_state(28,2013)}

\dontrun{
fars_map_state(Alabama,2013)# error
}
}
