% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{\code{fars_map_state} will plot the accidents on a map via the maps package. As inputs we have
the state number and the year of interest}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number}

\item{years}{The year of interest}
}
\description{
\code{fars_map_state} will plot the accidents on a map via the maps package. As inputs we have
the state number and the year of interest
}
\examples{
fars_map_state(12, 2014)

}

