% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents on a map for a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer code for the state of interest}

\item{year}{integer year of interest}
}
\value{
A map of the state with the accidents plotted as points
}
\description{
This function plots the location of accidents on a map for a given state and year, using the Federal
Highway Administration's Fatality Analysis Reporting System (FARS) data.
}
\examples{
\dontrun{
# plot accidents for state number 10 in year 2014
fars_map_state(10, 2014)
}
}
