% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the Accidents Data.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer denoting the state to plot for.}

\item{year}{The year for which the plot is required.}
}
\value{
Plot of the cleaned data on a US map.
}
\description{
The function takes a particular state and year into account and plots the corresponding locations
             of casualties on a map of USA. The `state.num` feature of the data ranges from 1 to 56.
}
\note{
An error is raised if the state idenification number (`state.num`) is invalid.

No plot is returned if number of accidents is zero in the specified month and year.
}
\examples{
\dontrun{
fars_map_state(48, 2013)
fars_map_state(56, 2014)
fars_map_state(82, 2013) # error is raised as no data for the specified state number exists.
}

}
