% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Creates a geographical plot showing the accidents for a given year in a given
   state. Each accident is plotted as a single dot.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number which you would like displayed.}

\item{year}{The year that you would like in the filename.}
}
\value{
Displays a plot (no object returned).
}
\description{
Creates a geographical plot showing the accidents for a given year in a given
   state. Each accident is plotted as a single dot.
}
\note{
If the state.num argument doesn't correspond to a state within the 
   specified data frame's STATE column, and error will be thrown.

If there are no accidents to plot for the specified state.num/year combination
  a message will display saying so.
}
