% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Mapping FARS data for a specific state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number between 1 and 56 representing the states in USA}

\item{year}{integer, the year for which the data is plotted}
}
\value{
A plot of the state and dots representing the places where the accidents happened. If state.num representing state number is not valid, we stop. Furthermore,
if for one state there is not data, this means there are no accidents
}
\description{
The function shows FARS data on a map for a specific state and year.
If the state number is invalid, the error message is shown.
}
\examples{
fars_map_state(1,2013)
\dontrun{
fars_map_state(1,2013)
#' }
}
