% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Reporting System data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string containing filename}
}
\value{
This function returns a tibble with data from selected file.
}
\description{
The function fars_read() enables reading FARS files into R.
}
\note{
The function will return an error if a file doesn't exist.
}
\examples{
\dontrun{
my.data <- fars_read('accident_2015.csv.bz2')
}

}
