% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file with FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the text the function will print}
}
\value{
This function returns a data frame as a tibble
}
\description{
This is a function that reads the data from a given file abd returns the data as a tibble. If the file does not exist it stops
}
\examples{
\dontrun{
data<-fars_read("data.csv")
data<-fars_read(filename="data.csv")
}

}
