% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character vector of the name of a .csv file to load}
}
\value{
Data frame containing the contents of the file named in \code{filename}
}
\description{
\code{fars_read} Returns a data frame from a file.
}
\details{
This function calls read_csv from the readr library to read in a csv file
and uses tbl_df to convert the tibble to a data frame.
\code{frs_read()} will fail if the file does not exist.
}
\examples{
\dontrun{
fars_read("./my_csv_file.csv")
}
}

