% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsdata.R
\name{fars_read}
\alias{fars_read}
\title{Read the FARS data.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the filename of the file as a string.}
}
\value{
the contents of the file as a tibble.
}
\description{
Read the file and convert it to a tibble.
}
