% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path to the file to be read}
}
\value{
if the file exists at the path, it returns a data frame
    with the parsed data using the column types inferred
    from the first 1000 rows of data. If the file name does not exist
    the function stops with a message
}
\description{
Function that reads in the fars data
}
\details{
The function uses read_csv to import the data and then converts it to a dataframe
    using tbl_df.  read_csv is silent through the use of suppressMessages

    stops processing if the file does not exist at the path
}
\examples{
fars_read("accident_2013.csv.bz2")

}
