% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Function to read the fars csv file data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{is a string specifying the address where to read the fars cdv (zipped) data}
}
\value{
data are the loaded fars data in dataframe format
}
\description{
This function checks whether the given filename is correct and then reads the fars csv file data
}
\details{
This function first checks whether the filename is correct: if this is not correct the execution of the function is
stopped. Then the csv data is loaded using the read_csv function from the readr package (any messages are suppressed)
and then transformed into a dataframe using the tbl_df function from the dplyr package. The dataframe has the following columns:
\describe{
\item{STATE}{State number}
\item{ST_CASE}{State/Case Number}
\item{VE_TOTAL}{Number of Vehicles in Crash}
\item{VE_FORMS}{Vehicle Forms Submitted}
\item{PVH_INVL}{Number of Parked/Working Vehicles in the Crash}
\item{PEDS}{Number of Persons Not in Motor Vehicles}
\item{PERNOTMVIT}{Number of Persons Not in Motor Vehicles in Transport}
\item{PERMVIT}{Number of Persons in Motor Vehicles In-Transport}
\item{PERSONS}{Number of Forms Submitted for Persons in Motor Vehicles}
\item{COUNTY}{County location}
\item{CITY}{City location}
\item{DAY}{Day of the Month of the Crash}
\item{MONTH}{Month of Crash}
\item{YEAR}{Year of Crash}
\item{DAY_WEEK}{Day of Week of crash}
\item{HOUR}{Hour of Crash}
\item{MINUTE}{Minute of Crash}
\item{NHS}{Did this crash occurr on a trafficway that is part of the National Highway System?}
\item{ROAD_FNC}{Functional classification of the trafficway on which the crash occurred}
\item{ROUTE}{Route signing of the trafficway on which the crash occurred}
\item{TWAY_ID}{This data element records the trafficway on which the crash occurred (1982-Later)}
\item{TWAY_ID2}{Trafficway on which the crash occurred (2004-Later)}
\item{MILEPT}{Milepoint nearest to the location where the crash occurred}
\item{LATITUDE}{Location of the crash using Global Position coordinates: latitude}
\item{LONGITUD}{Location of the crash using Global Position coordinates:: longitude}
\item{SP_JUR}{Does the location on the trafficway where the crash occurred qualify as a Special Jurisdiction?}
\item{HARM_EV}{First injury or damage producing event of the crash}
\item{MAN_COLL}{Orientation of two motor vehicles in-transport when they are involved in the "First Harmful Event" of a collision crash}
\item{RELJCT1}{Crash's location with respect to presence in an interchange area}
\item{RELJCT2}{Crash's location with respect to presence in or proximity to components typically in junction or interchange areas}
\item{TYP_INT}{Type of Intersection}
\item{WRK_ZONE}{Crash  within the boundaries of a work zone}
\item{REL_ROAD}{Location of the crash as it relates to its position within or outside the trafficway}
\item{LGT_COND}{Light Condition}
\item{WEATHER1}{Atmospheric Conditions (2007-Later)}
\item{WEATHER2}{Atmospheric Conditions (2007-Later)}
\item{WEATHER}{Atmospheric Conditions (1975-2006)}
\item{SCH_BUS}{School Bus Related to crash}
\item{RAIL}{Did the crash occurr in or near a rail grade crossing}
\item{NOT_HOUR}{Hour that emergency medical service was notified}
\item{NOT_MIN}{Minutes after the hour that emergency medical service was notified}
\item{ARR_HOUR}{Hour of Arrival at Scene}
\item{ARR_MIN}{Minute of Arrival at Scene}
\item{HOSP_HR}{Hour of EMS Arrival at Hospital}
\item{HOSP_MN}{Minute of EMS Arrival at Hospital}
\item{CF1}{factors related to the crash expressed by the investigating officer}
\item{CF2}{factors related to the crash expressed by the investigating officer}
\item{CF3}{factors related to the crash expressed by the investigating officer}
\item{FATALS}{number of fatally injured persons}
\item{DRUNK_DR}{number of drunk drivers involved in the fatal crash}
}
}
\examples{
 data_file<-system.file("extdata","accident_2013.csv.bz2", package = "farsdata")
 setwd(dirname(data_file))
 file="accident_2013.csv.bz2"
 dat<-fars_read(file)

 \dontrun{
 file="accident_2011.csv.bz2"
 dat<-fars_read(file) # error
 }
}

