% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Import CSV into a data frame tbl.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of the file which the data are to be read from}
}
\value{
A data frame (tbl) containing a representation of the data in the file.
}
\description{
This is a simple function that, by default, is useful for reading the fars data
the fars is a data set from the US National Highway Traffic Safety Administration's Fatality Analysis
Reporting System, which is a nationwide census providing the American public yearly data regarding
fatal injuries suffered in motor vehicle traffic crashes
}
\examples{
\dontrun{
fars_read("accident_2013.csv")
}

}
