% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The full file path to the CSV file containing the fars data.}
}
\value{
tibble object containing the data from the CSV file
}
\description{
fars_read checks for CSV file, returns error or loads file
}
\details{
This function checks for the specified CSV file using the file name provided. If the file does not exist,
an error is returned stating that the file does not exist. If the file does exist, the file is loaded as a data frame.
}
\examples{
\dontrun{
data <- fars_read("accident_2013.csv.bz2")
head(data)
}
}
