% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Title: fars_read
The function fars_read reads the file used as argument. The input file has to be in csv format and the function returns a dataframe of type tibble.
If the file does not exist an error will be printed.}
\source{
fars_data/accident_year.csv.bz2
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{(a csv file)}
}
\value{
a dplyr tbl_df (a tibble dataframe)
}
\description{
Title: fars_read
The function fars_read reads the file used as argument. The input file has to be in csv format and the function returns a dataframe of type tibble.
If the file does not exist an error will be printed.
}
\examples{
\dontrun{x <- fars_read('myFile.csv')}
\dontrun{system.file("fars_data", "accident_year.csv.bz2", package = "packFars"). }

}
