% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{----1-----
fars_read: Reads a csv file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the csv file.}
}
\value{
The function returns a tibble based on the specified csv file.
}
\description{
The function reads a csv file defined by \code{filename} argument and returns
             a tibble. If the path is incorrect the function will end and return an error.
}
\examples{

\dontrun{
tb_accident_2015 <- fars_read("./data/accident_2015.csv.bz2")
}

}

