% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading Fars Data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the file name of a .csv file that you wish to read into the data}
}
\value{
A tibble data frame object containing the FARS dara.
}
\description{
\code{fars_read} is a function that reads in Fatality Analysis Reporting System data.
 You can select the file wanted (using the \code{filename} argument)
}
\note{
If the specified filename does not exist,
 returns "file '" \code{filename}"' does not exist".
}
\examples{
fars_read("accident_2013.csv.bz2")

\dontrun{
fars_read("doesnt_exist.csv") # error
}
}
