% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in the "Fatality Analysis Reporting System" data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{data file with csv format}
}
\value{
\code{fars_read} looks for the filename within the current path directory. The function stops, when
it doesn't find the specified file. If the file exists, the function loads the csv file and returns a data frame
(side note: it ensures that it is a data frame via a dplyr function)
}
\description{
\code{fars_read} reads the data file from the "Fatality Analysis Reporting System.
}
\examples{
full_filename <- system.file('extdata', 'accident_2013.csv.bz2', package = 'farsdata')

}

