% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{path to CSV file to be read (compressed CSV files will be
uncompressed automatically)}
}
\value{
a tibble containing data
}
\description{
This function is used to import Fatality Analysis Reporting System (FARS) data.
It is not specific to FARS. Error will be thrown if file does not exist.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
fars_read("accident_2014.csv.bz2")
}

}
