% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data from a CSV file and return a tibble
This function reads FARS data from CSV files using the function \code{read_csv} from
the readr package and then stores it as a tibble using the function \code{tbl_df}
from the dplyr package}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{This is the file name of the FARS data to be read if the data is in the current
working directory otherwise it should be the file path of the data}
}
\value{
A tibble containing the data from the CSV file
}
\description{
Read FARS data from a CSV file and return a tibble
This function reads FARS data from CSV files using the function \code{read_csv} from
the readr package and then stores it as a tibble using the function \code{tbl_df}
from the dplyr package
}
\note{
fars_read depends on read_csv and tbl_df from the readr and dplyr packages respectively
}
\examples{
\dontrun{
fars_read("accident_2013.csv")
fars_read("path/to/your/csv/file.csv")
}
}
