% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read a Accidents Data File.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The filename of the dataset to be imported.}
}
\value{
A dataframe or object of class `tbl_df`.
}
\description{
The data files feature US National Highway Traffic Safety Administration's Fatality Analysis
             Reporting System, which is a nationwide census providing the American public yearly data
             regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\note{
An error is thrown if a file with the given filename doesn't exists.
}
\examples{
\dontrun{

fars_read(accident_2015.csv.bz2)

fars_read(accident_2020.csv.bz2) # shows an error as such a file doesn't exists.

}

}
