% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a CSV file into a tibble (tidyverse dataframe)}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The aboslute or relative path to the filename to read into
R (should be a CSV file) .}
}
\value{
The CSV data in tibble (tidyverse version of data frames) format.
}
\description{
This function reads a CSV file into R as a tibble.
}
\note{
If the file doesn't exist, a message will be displayed saying so.
}
\examples{
df <- fars_read("./data/data.csv")
df <- fars_read("./data/accident_2013.csv.bz2")
}
