% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character string, ath to the csv file}
}
\value{
Returns a tibble with the lodaded FARS data.
}
\description{
The function reads FARS (Fatality Analysis Reporting System) data from a csv file.
If the file does not exist, an error message is shown.
}
\examples{
fars_read("accident_2013.csv.bz2")
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
