% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reading years and months from Fatality Analysis Reporting System data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years to read the datasets for}
}
\value{
Function returns a list of tibbles for each selected year, containing month and year columns.
}
\description{
The function reads years and months from multiple FARS files.
}
\note{
The files should be named 'accident_year.csv.bz2' otherwise the function will return an error.
}
\examples{
my.data.lst <- fars_read_years(c(2013, 2014, 2015))
my.data.lst <- fars_read_years(2015)

}
