% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Fars years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector with a list of years}
}
\value{
The function returns a data frame including entries in data by month or NULL if the year is not valid
}
\description{
Help function used fars_summarize_years
}
\examples{
\dontrun{
fars_read_years(2013)
}

}
