% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsdata.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the data for a list of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{the list or vector of integer years to read.}
}
\value{
a list with a tibble for each year containing columns
  MONTH and year. If the file does not exist or does not have
  the correct data, then a warning is emitted and NULL returned.
}
\description{
Successively read the accidents files for each
year provided and collects the tibbles in a list.
Each tibble contains only year and month columns.
}
\examples{
\\notrun{
fars_read_years(c(2013, 2014))
}

}
