% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{list of years to return the data}
}
\value{
returns a list of dataframes that contain year and month information.
    The length of the list is the same as the number of items in the
    input parameter years.  Invalid years or years with no data are returned
    as NULL with an error message
}
\description{
generates the year and month data for the years in question
}
\details{
The function loops over all the elements of the "years" and extracts a data frame
    consisting of the "MONTH" column from the original data file and
    a computed "year" column based on the current element of the "years" param
}
\examples{
\dontrun{
fars_read_years(c("2013","2014","2015"))
fars_read_years(c(2013,2014))
}

}
