% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Function to read the fars csv file data over several years extracting the months of the crashes}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{is a numeric vector specifying which yearly data
(regarding fatal injuries suffered in motor vehicle traffic crashes)
have to be loaded}
}
\value{
a list containing the loaded fars data in dataframe format for each year
}
\description{
This function checks whether the given years are correct,
then reads the fars csv file data and extracts the months of the crashes.
If the years are invalid it sends a warning
}
\details{
This function checks whether the given years are correct,
then reads the fars csv file data and extracts the months of the crashes.
If the years are invalid it sends
a warning. Remember that only 3 years are available: 2013,2014,2015
}
\examples{

 data_file<-system.file("extdata","accident_2013.csv.bz2", package = "farsdata")
 setwd(dirname(data_file))
 dat=fars_read_years(c(2013,2014))
 dat

}

