% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{reading fars data according to years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integer indicates the year of the fars file}
}
\value{
a list of data frame(tibble), each component of the list contains a tibble with two columns
   of MONTH and year of the accident
}
\description{
This function read the fars data of a specific year and select only month and year of the accident
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014))
}

}
