% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of four-digit years representing the data of interest}
}
\value{
a list of tibble objects combining the month and year of accidents from each year provided, or NULL for invalid years
}
\description{
fars_read_years takes a list of years, builds the file names for each year provided, and loads the files
}
\details{
This function takes a list of four-digits numeric years, creates a list of file names using the years provided,
and checks for the specified CSV file using the file names created. If the file does not exist,
an error is returned stating that the an invalid year is provided and NULL is returned. If the file does exist, the file
is loaded as a data frame containing the month and year with one row for each accident that occurred.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014))
}
}
\seealso{
\link{fars_read}

\link{make_filename}
}
