% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Title: fars_read_years
The function fars_read_years accepts a list of one or more years and calls the function make_filename() with the years. Then, it fills the files with the data of its specific year.
Uses make_filename(year)
fars_read(file)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{(a list with the years we want to obtain the data from). If the year is not valid, an error message will appear.}

\item{year}{(the year from which we want the data file).}

\item{e}{(error).}
}
\value{
one or more datasets depending on the year/s given. If there is an error it returns NULL.
}
\description{
Title: fars_read_years
The function fars_read_years accepts a list of one or more years and calls the function make_filename() with the years. Then, it fills the files with the data of its specific year.
Uses make_filename(year)
fars_read(file)
}
\examples{
\dontrun{fars_read_years(2013)}
\dontrun{fars_read_years(as.list(2013, 2015, 2016)))}
\dontrun{fars_read_years(2013:2016)}

}
