% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS Files over multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of strings, each containing a number (integer, numeric or string) that
can be coerced into a string, to represent the years of the data required}
}
\value{
A list of data frames, each containing the month and year data for the Fars data sets requested.
}
\description{
\code{fars_read_years} returns a list of data frames containing Fatality Analysis Reporting System data,
 each data frame containing the year and month data of observations, but not the obsevations themselves
}
\note{
if an invalid year is given, returns \code"invalid year"
}
\examples{
fars_read_years(2013)
fars_read_years(c(2013,2014))
}
