% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{\code{fars_read_years} reads multiple files in, based on the data given by the "Fatality
Analysis Reporting System"}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The years relating to the file names, to be read in}
}
\value{
\code{fars_read_years} will search for the file names based on the input \code{years}
For example, if \code{years} is specified as \code{2016:2017}, the function will search for
the following files:
  \itemize{
    \item "accident_2016.csv.bz2"
    \item "accident_2017.csv.bz2"
  }
  If the files exist: a list containing the requested data will be returned.
  If the files do not exist: an error will be returned stating the invalid year(s).
}
\description{
\code{fars_read_years} reads multiple files in, based on the data given by the "Fatality
Analysis Reporting System"
}
\examples{
fars_read_years(2015:2016)

}
\seealso{
\code{\link{make_filename}} for naming convention
}
