% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data of multiple years}
\usage{
fars_read_years(years, fname_transform = identity)
}
\arguments{
\item{years}{a vector of years (will be coerced to integer)}

\item{fname_transform}{a function to be applied to filename that \code{make_filename}
contructs, default is \code{identity} which does not change filename. This is helpful
in cases where you want to add custom path prefix to all filenames.}
}
\value{
list of dataframes containing data for each year. names of the list
will be years and their corresponding values will have data for that year.
}
\description{
This function is for reading FARS data of multiple years. You
can provide a vector of years which will be converted to appropriate
file name using \link{make_filename}. After that, function will attempted to read
each file name. If filename does not exist, warning will be shown and \code{NULL}
will be returned for that years data.
}
\details{
If data for a year does not exist, returning list will have
\code{NULL} value associated with it
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}

}
