% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data for a list of years and return a tibble}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of integers specifying the years for which to read FARS data}
}
\value{
A tibble containing the number of accidents by month and year for all the specified years
}
\description{
This function reads FARS data from CSV files for a list of years using the fars_read function and
returns a tibble.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014, 2015))
}
}
