% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract Month and Year of Accidents Data.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The list or vector of years.}
}
\value{
An object of class `data.frame` or `tibble` or `tbl_df`.
}
\description{
The function actually takes data of all the concerned years into account and extracts the
             month and year of each accident, collects them and stores them in a dataframe
}
\note{
An error message is displayed if there is no data file corressponding to some input year.
}
\examples{
\dontrun{

fars_read_years(list(2013, 2014, 2015))  # Input is a list

fars_read_years(c(2013, 2014, 2015))   # Input is a vector or a list

}

}
