% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{\code{fars_read_years} : Reads the year and finds the csv file matching
the same year. If the csv file is not found, 'invalid year' error is displayed.
if the year is appropriate,  the csv file is read and mutate a new column with
the year and then select the columns MONTH and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of year to be searched}
}
\value{
A dataset with of the relevant year, with the columns MONTH and year.
Otherwise, an error 'invalid year' is displayed.
}
\description{
\code{fars_read_years} : Reads the year and finds the csv file matching
the same year. If the csv file is not found, 'invalid year' error is displayed.
if the year is appropriate,  the csv file is read and mutate a new column with
the year and then select the columns MONTH and year
}
