% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reading FARS data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{integer, a vector of years}
}
\value{
returns a tibble object with FARS data for specified years by month. Imported file is imported via function fars_read. If the year is not existing in the data
then warning gets created stating "invalid year"
}
\description{
The function reads FARS data for multiple years and combines them into a tibble.
If the year is not valid, the warning message is shown.
}
\examples{
far_read(fars_read_years)
\dontrun{
far_read(fars_read_years)
}

}
