% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise Fatality Analysis Reporting System data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years to read the datasets for}
}
\value{
The function returns a tibble with number of cases.
}
\description{
This function calculates the number of accidents per month in selected years.
}
\note{
The files should be named 'accident_year.csv.bz2' otherwise the function will return an error.
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
}

}
