% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A character or numeric vector of target \code{years}}
}
\value{
A data frame
}
\description{
Summary data by year and month from a list of target years.
}
\details{
Loads the FARS files associated with each year, then summarizes by each
year and month in the data.
Data files must be in .bz format and must have the name format of
"accident_year.csv.bz2" for each year in the \code{years} vector.
}
\examples{
\dontrun{
fars_summarize_years(c("2016", "2017"))
fars_summarize_years(1980:1990)
}
}

