% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsdata.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Generate monthly accidents summary table.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{the list of integer years to read.}
}
\value{
a tibble containing the monthly totals for each year.
}
\description{
Read the data for some years and generates
a table containing the monthly totals for
each year.
}
\examples{
\\notrun{
fars_summarize_years(c(2013, 2014, 2015))
}

}
