% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{list of years to return the data}
}
\value{
a dataframe that contain up to 12 rows of data with one column
    of summary data per input year
}
\description{
generates 12-by-n monthly summary data for the n-years in question
}
\details{
The function binds all the rows returned in the yearly dataframes by
    fars_read_years and computes the monthly summaries and
    converts the summary into a 12 row monthly dataframe with
    one column per row

    the dataframe may have less than 12 rows if the summary data doesn't
    exist for any particular row
}
\examples{
fars_summarize_years(c("2013","2014","2015"))
fars_summarize_years(c(2013,2014))

}
