% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function to read the fars csv file data over several years,
summaryzing the crashes for each month of the given years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{is a numeric vector specifying which yearly data
(regarding fatal injuries suffered in motor vehicle traffic crashes)
have to be loaded}
}
\value{
a dataframe containing the number of crashes for each month and each year
}
\description{
This function checks whether the given years are correct,
then reads the fars csv file data, extracts the months of the crashes,
compute the number of crashes for each month, and then prepare a dataframe
where the rows are the months, while the columns are the year
}
\details{
This function loadsthe fars data and summarize then by preparing a dataframe
containing the number of crashes for each month (rows) for each given year (columns)
}
\examples{

 data_file<-system.file("extdata","accident_2013.csv.bz2", package = "farsdata")
 setwd(dirname(data_file))
 dat=fars_summarize_years(c(2013,2014))
 dat

}

