% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars data summary}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of integer indicates the year of the fars file}
}
\value{
a data frame(tibble) with 12 rows and one column for each year
}
\description{
This function read the fars data of a specific year and summarize the data in a tibble,
  this data frame has 12 rows for month, and one column for each year. In each cell you can
  find the number of the accident in that YEAR, MONTH.
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014))
}

}
