% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of four-digit years representing the data of interest}
}
\value{
tibble object of accident data summarized by Month and Year
}
\description{
fars_summarize_years takes a list of years and summarizes the number of accidents by month and year
}
\details{
This function uses the four-digit numeric years provided to check the CSV file associated with the specified years. If the
file does not exist, an error is returned stating that the file does not exist. If the file does exist, the file is loaded
into a list of data frames. The lists are then combined into one data frame, grouped by month and years of data, and summarized
by counting the total number of accidents for each month and year combination.
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014))
}
}
\seealso{
\link{fars_read_years}
}
