% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Title: fars_summarize_years
The function fars_summarize_years accepts a list of one or more years and passes the list to fars_read_years().
Then it receives from fars_read_years() a data set in columns month and year, and counts the number of observations for a year by month.
Uses fars_read_years(years)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{(a list with the years we want to obtain the data from).}
}
\value{
a dataframe with the relative data of the years introduced (counts by month and year).
}
\description{
Title: fars_summarize_years
The function fars_summarize_years accepts a list of one or more years and passes the list to fars_read_years().
Then it receives from fars_read_years() a data set in columns month and year, and counts the number of observations for a year by month.
Uses fars_read_years(years)
}
\examples{
\dontrun{fars_summarize_years(2013)}
\dontrun{fars_summarize_years(as.list(2013, 2015, 2016)))}
\dontrun{fars_summarize_years(2013:2016)}

}
