% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{----4-----
fars_summarize_years: Counts number of accidents per month & year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of years (numeric or integer) that will be
searched in the data}
}
\value{
Returns a tibble in a wide format (pivot) with months in rows and defined
        years in columns, using as values the number of accidents. Returns a warning for
        every input year that does not exist in the datasets.
}
\description{
Based on the user-selected years, the function summarises the number of accidents
             in the US on a monthly basis. The accident files need to be in the working
             directory.
}
\examples{
\dontrun{
fars_summarize_years(2015:2016)
}

}

