% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS Summary by Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of strings, each containing a number (integer, numeric or string) that
can be coerced into a string, to represent the years of the data required}
}
\value{
Returns dataframe of month by year values for FARS data, will only return years with valid FARS data
}
\description{
\code{fars_summarize_years} presents a summary of
 Fatality Analysis Reporting System data by year and month
 over a given vector of years
}
\examples{
fars_summarise_years(c(2013,2014,2015))
\dontrun{
fars_summarise_years("3009")# error
}
}
