% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{\code{fars_summarize_years} summarizes the data (given by the "Fatality Analysis Reporting System")
according to the years, which is being defined.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years to be summarized}
}
\value{
\code{fars:summarize_years} gives the summarized data in a wide format
}
\description{
\code{fars_summarize_years} summarizes the data (given by the "Fatality Analysis Reporting System")
according to the years, which is being defined.
}
\examples{
fars_summarize_years(2015:2016)

}
\seealso{
\code{\link{fars_read_years}} to get more information
}
