% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarise FARS data}
\usage{
fars_summarize_years(years, ...)
}
\arguments{
\item{years}{a vector of years (will be coerced to integer)}

\item{...}{further arguments to be passed to \code{fars_read_years}}
}
\value{
tibble having summary metrics
}
\description{
This function derives summary metrics for every year and month
for FARS data. First it imports data using \link{fars_read_years} and
then produces a pivot table showing row count for every year and month.
Months are in rows and years are in column.
}
\examples{
\dontrun{
fars_summarize_years(2014)
fars_summarize_years(2013:2015)
}

}
