% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data for a list of years and return a tibble}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of integers specifying the years for which to summarize FARS data}
}
\value{
A tibble containing the number of accidents by month and year for all the specified years,
spread across multiple columns
}
\description{
This function reads FARS data from CSV files for a list of years using the fars_read_years function and
returns a tibble containing the number of accidents by month and year, spread across multiple columns.
}
\note{
This functions depends on the bind_rows, group_by, summarize functions from dplyr and spread from tidyr
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014, 2015))
}
}
