% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the Data by Month and Year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of years.}
}
\value{
An object of class `data.frame` or `tibble` or `tbl_df`.
}
\description{
A function to show the distribution of fatalities in motor vehicle accidents over month and year
             of occurrence in USA in a tabulated format.
}
\examples{
\dontrun{

fars_summarize_years(list(2013, 2014, 2015))  # Input is a list

fars_summarize_years(c(2013, 2014, 2015))   # Input is a vector or a list

fars_summarize_years(list(2013, 2014, 2020)) # Error raised as 2020 is not a valid year for input.

}

}
