% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a table summarizing the count of fatal car incidents in a given month
and year combination.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years for which the number of fatal car incidents
should be displayed, grouped by Month.}
}
\value{
a data frame.
}
\description{
Creates a table summarizing the count of fatal car incidents in a given month
and year combination.
}
\note{
If one of the years specified doesn't have an associated file, and error
   will be thrown.
}
