% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create filename for Fatality Analysis Reporting System data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year of the dataset}
}
\value{
Function returns a character string with filename including year of the dataset.
}
\description{
This is a helper hunction to create a filename including a year of the dataset.
}
\examples{
make_filename(2017)

}
