% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character or numeric vector of the desired year(s)}
}
\value{
A character vector containing FARS filenames for given years
}
\description{
Creates filenames for loading FARS files from a given year
}
\details{
Takes a character or numeric vector of target years and returns a character
vector of FARS filenames for the target years in the format \code{"accident_year.csv.bz2"}
}
\examples{
make_filename("2017")
make_filename(2017)
make_filename(c("2004", "2005", "2006"))
make_filename(2010:2016)
}

