% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsdata.R
\name{make_filename}
\alias{make_filename}
\title{Construct the filename for the year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year for which to construct the filename.
Can be an integer or a string that can be converted to
an integer.}
}
\value{
the constructed filename as a string.
}
\description{
Construct the filename of the accidents data
file in csv.bz2 format for the given year.
}
