% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{param with the year information
will be coerced to a integer}
}
\value{
string of the form "accident_yyyy.csv.bz2" where yyyy repersents the value
    passed in the "year" param
}
\description{
create a data file name given an input year
}
\details{
The function ettempts to explicitly convert the year param to
    an integer.  This may return NA resulting in "accident_NA.csv.bz2"
    as the returned file name as well as a warning message

   #'
}
\examples{
\dontrun{
make_filename("2013")
make_filename("2013a")
}

}
