% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create the filename for the fars csv file data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{is a numeric scalar specifying which yearly data
(regarding fatal injuries suffered in motor vehicle traffic crashes) have to be loaded}
}
\value{
a string specifying the address where to read the fars data
}
\description{
This function creates the filename for the fars csv (zipped) file data
}
\details{
This function  creates the filename for the specific yearly fars csv (zipped) file data.
There are 3 possible years: 2013, 2014, 2015.
}
\examples{
 file=make_filename(2013)
 file
}

