% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{fars file name creation}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer indicates the year of the fars file}
}
\value{
a string which indicates the file name of the fars data
}
\description{
Making the file name of a fars data according to the year of occurence
}
\examples{
\dontrun{
make_filename(2013)
}

}
