% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The four-digit numeric year of data for which to create the file name.}
}
\value{
a string of the file name based on the year: "accident_xxxx.csv.bz2", where xxxx represents the year
}
\description{
make_filename creates a CSV file name using the 4-digit Year provided as a parameter (2013, 2014, or 2015)
}
\details{
This function creates a file name based on the four-digit year provided to the function.
}
\examples{
\dontrun{
make_filename("2013")
}
}
