% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Title: make_filename
The function make_filename creates data files by year in bz2 format. It transforms the argument year into an interger and sends the determined name of the file to a string.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{(the year from which we want the data file).}
}
\value{
a filename of the kind 'accident_year.csv.bz2'.
}
\description{
Title: make_filename
The function make_filename creates data files by year in bz2 format. It transforms the argument year into an interger and sends the determined name of the file to a string.
}
\examples{
\dontrun{x <- make_filename("2013")}
\dontrun{system.file("fars_data", "accident_year.csv.bz2", package = "packFars")}

}
