% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make FARS Filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number (integer, numeric or string) that can be coerced into a string to represent the year of the data,
can be any number.}
}
\value{
The function will return a string of the filename in propper FARS data form, for the given year.
}
\description{
\code{make_filename} formats a filename for Fatality Analysis Reporting System data, given a year as input
}
\examples{
make_filename(year = '2013')
make_filename(year = 2013)

\dontrun{
make_filename(year="twenty eighteen") # error
}
}
