% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates the string for a filename, according to a given year. Only valid for the "Fatality
Analysis Reporting System", respectively their files}
\usage{
make_filename(year)
}
\arguments{
\item{year}{input year, can be numeric or character}
}
\value{
\code{make_filename} gives you back the file name as a string
}
\description{
\code{make_filename} takes as input a year and gives out the name of the file as a string
}
\examples{
make_filename(2015)

}

