% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make FARS data filename from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year of FARS data (coerced to integer)}
}
\value{
a FARS data file name corresponding to given year
}
\description{
This functions creates a full FARS data file name given a year
}
\examples{
make_filename(2002)
make_filename(1995)

}
