% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate a file name for FARS data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer specifying the year for which to generate a file name}
}
\value{
A character string containing the file name for the FARS data for the given year
}
\description{
This function generates a file name for FARS data based on a given year.
}
\examples{
make_filename(2013)

}
