% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Create a Filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for which the records are desired.}
}
\value{
A string that denotes the name of the data file.
}
\description{
A function that reproduces the name of the file that consists of the data on accidents for the given year.
}
\note{
It is recommended to take the input as 2013, 2014 or 2015 since data beyond these 3 years is unavailble.
}
\examples{
\dontrun{
make_filename(2013)  # Input is numeric
make_filename("2015")  # Input is a string
make_filename(as.double(2018))  # Input is of `double` data type.
}

}
