% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a descriptive filename based on the input year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year that you would like in the filename.}
}
\value{
A character object that represents the filename for car fatality data.
}
\description{
Creates a descriptive filename based on the input year.
}
