% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creating file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer, The year for which the filename is generated}
}
\value{
A character string representing the generated filename. Output is the file name of the form accident_year.csv.bz2, where year is specified by the user
}
\description{
The function generates the filename for a FARS data file, based on the input year.
}
\examples{
make_filename(2019)
\dontrun{
make_filename(2019)
}


}
